refactor: move some tools to dedicated files
Due to the new PHPStan tool, we exceeded the line-length limit of 100 lines per file.
