+++ src/enum/toolType.ts
@@ -1,0 +1,4 @@
+export enum ToolType {
+    LINTER = 'linter',
+    CODE_CHECK = 'code_check',
+}
